src/main/java/com/wsulima/battledev/exercises/Exercise_1.java
package com.wsulima.battledev.exercises;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;
import java.util.Scanner;

public class Exercise_1 {

	public static void main(String[] argv) throws Exception {
		System.out.println(calculate());
	}

	public static String calculate() throws Exception {
		return calculate(readInputLines());
	}

	private static String calculate(List<String> lines) throws Exception {
		int shortestStrawLength = Integer.MAX_VALUE;
		String badLuckGuyName = "";

		for (String lengthAndName : lines.subList(1,lines.size())) {
			int length = toNumber(split(lengthAndName)[1]);
			String name = split(lengthAndName)[0];
			if(length < shortestStrawLength) {
				badLuckGuyName = name;
				shortestStrawLength = length;
			}
		}
		return badLuckGuyName;
	}


	/* ****************************************************************
								UTILS
	**************************************************************** */

	private static List<String> readInputLines() {
		List<String> result = new ArrayList<>();
		Scanner sc = new Scanner(System.in);
		while (sc.hasNextLine()) {
			result.add(sc.nextLine().trim());
		}
		return result;
	}

	private static String[] split(String text) {
		return text.split(" ");
	}

	private static Integer toNumber(String text) {
		return Integer.parseInt(text);
	}

	private static List<Integer> toNumbers(List<String> numbers) {
		return numbers.stream().map(Integer::parseInt).collect(java.util.stream.Collectors.toList());
	}

	private static Integer[] toNumbers(String... numbers) {
		return Arrays.stream(numbers).map(Integer::parseInt).toArray(Integer[]::new);
	}


}
